- refactor pom.xml - remain package name
