+++ src/main/java/com/moneyforward/ffp/smt/converter/fieldConverter/FieldConverter.java
@@ -1,4 +1,4 @@
-package com.moneyforward.smt.converter.fieldConverter;
+package com.moneyforward.ffp.smt.converter.fieldConverter;
 
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.errors.DataException;
