+++ src/main/java/com/moneyforward/ffp/smt/converter/fieldConverter/Float64FieldConverter.java
@@ -14,21 +14,21 @@
  * limitations under the License.
  */
 
-package com.moneyforward.smt.converter.fieldConverter;
+package com.moneyforward.ffp.smt.converter.fieldConverter;
 
 import org.apache.kafka.connect.data.Schema;
-import org.bson.BsonInt32;
+import org.bson.BsonDouble;
 import org.bson.BsonValue;
 
-public class Int16FieldConverter extends FieldConverter {
+public class Float64FieldConverter extends FieldConverter {
 
-    public Int16FieldConverter() {
-        super(Schema.INT16_SCHEMA);
+    public Float64FieldConverter() {
+        super(Schema.FLOAT64_SCHEMA);
     }
 
     @Override
     public BsonValue toBsonValue(Object data) {
-        return new BsonInt32(((Short) data).intValue());
+        return new BsonDouble((Double) data);
     }
 
 }
